requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
Authlib==1.6.5
certifi==2025.10.5
cffi==2.0.0
click==8.3.0
cryptography==46.0.3
fastapi==0.120.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
python-dotenv==1.2.1
sniffio==1.3.1
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
